fixed Windows warnings, updated Windows test project

+++ test/general-loader/test-general-loader.cpp
@@ -198,7 +198,7 @@ public:
     void OpenStream_OneTableOneColumn ( const char* p_dbName, const char* p_tableName, const char* p_columnName, size_t p_elemBits )
     {   
         SetUpStream_OneTable( p_dbName, p_tableName ); 
-        m_source . NewColumnEvent ( DefaultColumnId, DefaultTableId, p_columnName, p_elemBits );
+        m_source . NewColumnEvent ( DefaultColumnId, DefaultTableId, p_columnName, ( uint32_t ) p_elemBits );
         m_source . OpenStreamEvent();
     }
